+++ packages/circularprogress/src/css/index.ts
@@ -4,7 +4,7 @@ import {
   names as themeNames
 } from '@pluralsight/ps-design-system-theme'
 
-import * as vars from '../vars'
+import * as vars from '../vars/index.js'
 
 export default {
   '@keyframes psds-circularprogress__keyframes__spin': {
