+++ packages/circularprogress/src/react/__stories__/index.story.tsx
@@ -2,7 +2,7 @@ import * as core from '@pluralsight/ps-design-system-core'
 import * as React from 'react'
 import { storiesOf } from '@storybook/react'
 
-import CircularProgress from '../..'
+import CircularProgress from '../../index.js'
 
 const valueStory = storiesOf('value', module)
 ;[0, 25, 50, 75, 100, 33, 66].forEach(value =>
