+++ packages/circularprogress/src/react/index.tsx
@@ -8,8 +8,8 @@ import {
 import { compose, css, keyframes, StyleAttribute } from 'glamor'
 import React from 'react'
 
-import stylesheet from '../css'
-import * as vars from '../vars'
+import stylesheet from '../css/index.js'
+import * as vars from '../vars/index.js'
 
 const radius = vars.style.width / 2 - vars.style.strokeWidth / 2
 const circumference = 2 * Math.PI * radius
